moved VavrConfig.java to config package
