+++ distributed-transactions/2-phase-commit/customer-service/src/main/java/org/example/customer/config/VavrConfig.java
@@ -1,4 +1,4 @@
-package org.example.order;
+package org.example.customer.config;
 
 import com.fasterxml.jackson.databind.Module;
 import io.vavr.jackson.datatype.VavrModule;
