+++ distributed-transactions/2-phase-commit/order-service/src/main/java/org/example/order/config/VavrConfig.java
@@ -1,4 +1,4 @@
-package org.example.customer;
+package org.example.order.config;
 
 import com.fasterxml.jackson.databind.Module;
 import io.vavr.jackson.datatype.VavrModule;
